% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcharts.R
\name{barcharts}
\alias{barcharts}
\title{Barcharts}
\usage{
barcharts(
  data,
  fill = "deepskyblue2",
  color = "grey30",
  labels = TRUE,
  sort = TRUE,
  maxcat = 20,
  abbrev = 20
)
}
\arguments{
\item{data}{data frame}

\item{fill}{fill color for bars}

\item{color}{color for bar labels}

\item{labels}{if \code{TRUE}, bars are labeled with percents}

\item{sort}{if \code{TRUE}, bars are sorted by frequency}

\item{maxcat}{numeric. barcharts with more than this number of bars will not be plotted.}

\item{abbrev}{numeric. abbreviate bar labels to at most, this character length.}
}
\value{
a ggplot graph
}
\description{
Create barcharts for all categorical variables in a data frame.
}
\examples{
barcharts(cars74)
}
