% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear.R
\name{clear}
\alias{clear}
\title{Clear/reset an RStudio session}
\usage{
clear(detach = TRUE)
}
\arguments{
\item{detach}{logical. If \code{TRUE}, unload
all but base packages.}
}
\description{
Clear/reset an RStudio session
}
\details{
The \code{clear()} function:
\enumerate{
\item Clears the console window
\item Deletes objects from the global environment
\item Deletes plots and plot history from the Plots window
\item Unloads all but base packages (unless \code{detach=FALSE})
\item Releases memory (garbage collection)
}
R scripts and Rmarkdown files are left untouched.
No objects saved to disk are affected.
}
\note{
This is a good way to start a fresh R session without
having to quit and restart RStudio. Additionally, any script
or Rmarkdown files remain open, the current working directory
is unchanged, and you remain in any open project.
}
\examples{
\dontrun{
clear()
}

}
