% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_plot.R
\name{mean_plot}
\alias{mean_plot}
\title{Mean plot with error bars}
\usage{
mean_plot(
  data,
  y,
  x,
  by,
  pointsize = 2,
  dodge = 0.2,
  lines = TRUE,
  width = 0.2,
  error = c("se", "sd", "ci"),
  ci = 0.95,
  robust = FALSE
)
}
\arguments{
\item{data}{a data frame.}

\item{y}{a numeric response variable.}

\item{x}{a categorical explanatory variable.}

\item{by}{a second categorical explanatory variable (optional).}

\item{pointsize}{numeric. Point size (default = 2).}

\item{dodge}{numeric. If a \code{by} variable is included, points and error bars are
dodged by this amount in order to avoid overlap (default = 0.2).}

\item{lines}{logical. If \code{TRUE}, group means are connected.}

\item{width}{numeric. Width of the horizontal top and bottom of error bars (default = 0.2). 
Set to 0 to produce pointranges instead of error bars.}

\item{error}{character. Each error bar represents either the standard deviation
\code{(sd)}, standard error of the mean  \code{(se)}, or confidence interval
\code{(ci)}.}

\item{ci}{numeric. if \code{error = "ci"}, this indicates the size of the confidence
interval. The default is \code{0.95} or a 95 percent confidence interval for the mean.}

\item{robust}{logical. If \code{TRUE}, the means, standard deviations, standard errors,
and confidence intervals are based on robust statistics. See \code{Details}. 
Default is \code{FALSE}.}
}
\value{
a \code{ggplot2} graph.
}
\description{
Plot group means with error bars. Error bars can represent standard deviations,
standard errors, or confidence intervals. Optionally, plots can be based on
robust statistics.
}
\details{
Robust statistics are based on deciles. The mean is estimated as the average of
the nine decile values dividing the numeric variable into 10 equal quantiles. The standard
deviation is bases on the sample standard deviation of the nine decile values. The standard 
error and confidence interval are calculated in the normal way, but use the robust
mean and standard deviation. See Abu-Shawiesh et al (2022).
}
\examples{
data(cars74)
mean_plot(cars74, mpg, cyl)
mean_plot(cars74, mpg, cyl, am)
mean_plot(cars74, mpg, cyl, am, error = "ci", 
          width = 0, lines = FALSE, robust = TRUE)
}
\references{
Ahmed Abu-Shawiesh, M., Sinsomboonthong, J., & Kibria, B. (2022). A modified robust
confidence interval for the population mean of distributrion baed on deciles. 
\emph{Statistics in Transition}, vol. \emph{23 (1)}.
\href{https://tinyurl.com/bdz7umj8}{pdf}
}
