% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize numeric variables}
\usage{
normalize(data, new_min = 0, new_max = 1)
}
\arguments{
\item{data}{a data frame.}

\item{new_min}{minimum for the transformed variables.}

\item{new_max}{maximum for the transformed variables.}
}
\value{
a data frame
}
\description{
Normalize the numeric variables in a data frame
}
\details{
\code{normalize} transforms all the numeric variables
in a data frame to have the same minimum and maximum values.
By default, this will be a minimum of 0 and maximum of 1.
Character variables and factors are left unchanged.
}
\note{
Use this function to be transform variables into a given range. The default
is [0, 1], but [-1, 1], [0, 100], or any other range is permissible.
}
\examples{
head(cars74)

cars74_st <- normalize(cars74)
head(cars74_st)
}
