% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phelp.R
\name{phelp}
\alias{phelp}
\title{Get help on a package}
\usage{
phelp(pckg)
}
\arguments{
\item{pckg}{The name of a package}
}
\value{
No return value, called for side effects.
}
\description{
\code{phelp} provides help on an installed package.
}
\details{
This function provides help on an installed package. The
package does not have to be loaded. The package name does
not need to be entered with quotes.
}
\examples{
phelp(stats)

}
