% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.crosstab.R
\name{plot.crosstab}
\alias{plot.crosstab}
\title{Plot a crosstab object}
\usage{
\method{plot}{crosstab}(x, size = 3.5, ...)
}
\arguments{
\item{x}{An object of class \code{crosstab}}

\item{size}{numeric. Size of bar text labels.}

\item{...}{no currently used.}
}
\value{
a ggplot2 graph
}
\description{
This function plots the results of a calculated two-way
frequency table.
}
\examples{
tbl <- crosstab(cars74, cyl, gear, type = "freq")
plot(tbl)

tbl <- crosstab(cars74, cyl, gear, type = "colpercent")
plot(tbl)
}
