% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.contents.R
\name{print.contents}
\alias{print.contents}
\title{Print a contents object}
\usage{
\method{print}{contents}(x, ...)
}
\arguments{
\item{x}{a object of class \code{contents}}

\item{...}{not used.}
}
\value{
No return value, called for side effects.
}
\description{
\code{print.contents} prints the results of the content function.
}
\examples{
testdata <- data.frame(height=c(4, 5, 3, 2, 100),
                       weight=c(39, 88, NA, 15, -2),
                       names=c("Bill","Dean", "Sam", NA, "Jane"),
                       race=c('b', 'w', 'w', 'o', 'b'))

x <- contents(testdata)
print(x)

}
