% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.crosstab.R
\name{print.crosstab}
\alias{print.crosstab}
\title{Print a crosstab object}
\usage{
\method{print}{crosstab}(x, ...)
}
\arguments{
\item{x}{An object of class \code{crosstab}}

\item{...}{not currently used.}
}
\value{
No return value, called for side effects
}
\description{
This function prints the results of a calculated two-way
frequency table.
}
\examples{
mycrosstab <- crosstab(mtcars, cyl, gear, type = "freq", digits = 2)
print(mycrosstab)

mycrosstab <- crosstab(mtcars, cyl, gear, type = "rowpercent", digits = 3)
print(mycrosstab)
}
