% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.tab.R
\name{print.tab}
\alias{print.tab}
\title{Print a tab object}
\usage{
\method{print}{tab}(x, ...)
}
\arguments{
\item{x}{An object of class \code{tab}}

\item{...}{Parameters passed to the print function}
}
\value{
No return value, called for side effects
}
\description{
Print the results of calculating a frequency table
}
\examples{
frequency <- tab(cardata, make, sort = TRUE, na.rm = FALSE)
print(frequency)
}
