% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qacEDA-package.R
\docType{package}
\name{qacBase-package}
\alias{qacBase}
\alias{qacBase-package}
\title{qacBase: Functions to Facilitate Exploratory Data Analysis}
\description{
Functions for descriptive statistics, data management, and data visualization.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rkabacoff/qacBase}
  \item Report bugs at \url{https://github.com/rkabacoff/qacBase/issues}
}

}
\author{
\strong{Maintainer}: Kabacoff Robert \email{rkabacoff@wesleyan.edu}

Other contributors:
\itemize{
  \item Barich Griffen [contributor]
  \item Jamrog Kelly [contributor]
  \item Kravchenko Elizaveta [contributor]
  \item Kuruvilla Jacob [contributor]
  \item Liu Lex [contributor]
  \item Nakamura Shota [contributor]
  \item Pham Kim [contributor]
  \item Rodriguez Belen [contributor]
  \item Ross Shane [contributor]
  \item Russo Chris [contributor]
  \item Corpuz Frederick [contributor]
  \item Juradat Nurah [contributor]
  \item Karp Harrison [contributor]
  \item Koech Kevin [contributor]
  \item Peters Anna [contributor]
  \item Shah Dhhyey [contributor]
  \item Stevenson Kenneth [contributor]
  \item Thomas-Franz Kaitlyn [contributor]
  \item Zheng Jiner [contributor]
  \item Aldarmaki Ahmed [contributor]
  \item Alneyadi Mohammed [contributor]
  \item Altai Chossis [contributor]
  \item Colorado Sofia [contributor]
  \item Northrop Blake [contributor]
  \item Peretz Shea [contributor]
  \item Qin Cher [contributor]
  \item Tuhabonye Emma [contributor]
  \item Wong Phillip [contributor]
}

}
\keyword{internal}
