% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcolors.R
\name{rcolors}
\alias{rcolors}
\title{R Colors}
\usage{
rcolors(color = NULL, cex = 0.6)
}
\arguments{
\item{color}{character. A text string used to search for specific color variations (see examples.)}

\item{cex}{numeric. text size for color labels.}
}
\value{
No return value, called for side effects
}
\description{
Plot a grid of R colors and their associated names
}
\details{
By default \code{rcolors} plots the basic 502 distinct colors provided by the
\code{colors} function. If a color name or part of a name is provided, only
colors with matching names are plotted.
}
\examples{
rcolors()
rcolors("blue")
rcolors("red")
rcolors("dark")

}
\references{
This function is adapted from code published by
\href{https://github.com/kbroman/broman}{Karl W. Broman}.
}
\seealso{
\link{colors}
}
