% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewness.R
\name{skewness}
\alias{skewness}
\title{Skewness}
\usage{
skewness(x, na.rm = TRUE)
}
\arguments{
\item{x}{numeric vector.}

\item{na.rm}{if \code{TRUE}, delete missing values.}
}
\value{
a number
}
\description{
Calculate the skewness of a numeric variable
}
\examples{
skewness(mtcars$mpg)
}
