% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_plot.R
\name{mean_plot}
\alias{mean_plot}
\title{Mean plot with error bars}
\usage{
mean_plot(
  data,
  y,
  x,
  by,
  pointsize = 2,
  dodge = 0.2,
  lines = TRUE,
  width = 0.2,
  error_type = c("se", "sd", "ci"),
  percent = 0.95,
  robust = FALSE
)
}
\arguments{
\item{data}{a data frame.}

\item{y}{a numeric response variable.}

\item{x}{a categorical explanatory variable.}

\item{by}{a second categorical explanatory variable (optional).}

\item{pointsize}{numeric. Point size (default = 2).}

\item{dodge}{numeric. If a \code{by} variable is included, points and error bars are
dodged by this amount in order to avoid overlap (default = 0.2).}

\item{lines}{logical. If \code{TRUE}, group means are connected.}

\item{width}{numeric. Width of the error bars (default = 0.2). 
Set to 0 to produce pointranges instead of error bars.}

\item{error_type}{character. Error bars represents either standard deviations
\code{(sd)}, standard errors of the means  \code{(se)}, or confidence intervals
\code{(ci)}. The default is the standard error.}

\item{percent}{numeric. if \code{error_type = "ci"}, this indicates the size of the confidence
interval. The default is \code{0.95} or a 95 percent confidence interval for the mean.}

\item{robust}{logical. If \code{TRUE}, the means, standard deviations, standard errors,
and confidence intervals are based on robust statistics. See \code{Details}. 
The default is \code{FALSE}.}
}
\value{
a \code{ggplot2} graph.
}
\description{
Plots group means with error bars. Error bars can be standard deviations,
standard errors, or confidence intervals. Optionally, plots can be based on
robust statistics.
}
\details{
Robust statistics are based on deciles, the nine values that divide the response variable into 
10 equal groups (where group containing roughly the same fraction of cases). The robust 
mean is the mean of these nine decile values. The robust standard deviation is the sample 
standard deviation of the nine decile values. The standard error and confidence interval are 
calculated in the normal way, but use the robust mean and standard deviation in 
their calculations. See Abu-Shawiesh et al (2022).
}
\examples{
mean_plot(cars74, mpg, cyl)
mean_plot(cars74, mpg, cyl, am)
mean_plot(cars74, mpg, cyl, am, 
          error_type = "ci", percent = 0.9,
          width = 0, lines = FALSE, robust = TRUE)
}
\references{
Ahmed Abu-Shawiesh, M., Sinsomboonthong, J., & Kibria, B. (2022). A modified robust
confidence interval for the population mean of distributrion baed on deciles. 
\emph{Statistics in Transition}, vol. \emph{23 (1)}.
\href{https://tinyurl.com/bdz7umj8}{pdf}
}
